Migrate conversion numba functions to conversion_numba_fns.py
